Adds much support for system commands.
Sorry for the code bomb. I was working around with the calls and parsing techniques in a sandbox project (which, unfortunately, I did not keep in version control because I thought it was going to be quick) and I got caught up in the details.

Moving the useful source from the sandbox to Nerdtool went fairly smoothly, and the outcome results in partial working. With a few bindings here and there, it should work as well as any of the other log types.