Just a bit of conversion to CoreData.
This is by no means complete, but it does (I believe) have a pretty good start on it. I am branching this in case I ever wanted to go back and pickup this thread. Right now, I don't feel that the move to CoreData is all that important or necessary.